+++ packages/storybook8/stories/Components/MessageThread/MessageThread.mdx
@@ -12,9 +12,11 @@ import { MessageThreadWithCustomMessagesExample } from './snippets/CustomMessage
 import { MessageThreadWithCustomMessageStatusIndicatorExample } from './snippets/CustomMessageStatusIndicator.snippet';
 import { MessageThreadWithCustomTimestampExample } from './snippets/CustomTimestamp.snippet';
 import { DefaultMessageThreadExample } from './snippets/Default.snippet';
-import { MessageThreadWithMessageStatusIndicatorExample } from './snippets/MessageStatusIndicator.snippet';
-import { MessageWithAttachment } from './snippets/MessageWithAttachment.snippet';
-import { MessageWithCustomAttachment } from './snippets/MessageWithCustomAttachment.snippet';
+import { MessageThreadWithMessageStatusIndicatorExample } from './snippets/MessageStatusIndicator.snippet';import {
+  MessageWithAttachment,
+  MessageWithAttachmentFromTeams,
+  MessageWithCustomAttachment
+} from './snippets/MessageWithAttachment.snippet';
 import { MessageWithCustomMentionRenderer } from './snippets/MessageWithCustomMentionRenderer.snippet';
 import { MessageThreadWithSystemMessagesExample } from './snippets/SystemMessages.snippet';
 import { MessageThreadWithInlineImageExample } from './snippets/WithInlineImageMessage.snippet';
@@ -31,7 +33,6 @@ import MessageThreadWithCustomTimestampExampleText from '!!raw-loader!./snippets
 import DefaultMessageThreadExampleText from '!!raw-loader!./snippets/Default.snippet.tsx';
 import MessageThreadWithMessageStatusIndicatorExampleText from '!!raw-loader!./snippets/MessageStatusIndicator.snippet.tsx';
 import MessageWithAttachmentText from '!!raw-loader!./snippets/MessageWithAttachment.snippet.tsx';
-import MessageWithCustomAttachmentText from '!!raw-loader!./snippets/MessageWithCustomAttachment.snippet.tsx';
 import MessageWithCustomMentionRendererText from '!!raw-loader!./snippets/MessageWithCustomMentionRenderer.snippet.tsx';
 import ExampleImportantsText from '!!raw-loader!./snippets/placeholdermessages.ts';
 import MessageThreadWithSystemMessagesExampleText from '!!raw-loader!./snippets/SystemMessages.snippet.tsx';
@@ -204,7 +205,7 @@ a static list for all secanrios.
 
 For example, the following code snippet demonstrates how to customize the download options for attachments.
 
-<Canvas mdxSource={MessageWithCustomAttachmentText}>
+<Canvas mdxSource={MessageWithAttachmentText}>
   <MessageWithCustomAttachment />
 </Canvas>{' '}
